requirements.txt
git+https://github.com/Rapptz/discord.py.git
openai
python-dotenv